% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP1805.R
\docType{data}
\name{CP1805.conc}
\alias{CP1805.conc}
\title{Pharmacokinetics of CP1805}
\format{A data frame with five variables:
\describe{
  \item{\code{ID}}{Subject identifier}
  \item{\code{DAY}}{Day of measurement}
  \item{\code{TIME.AFTER.DOSE}}{Time since the dose on the current \code{DAY} (in hours)}
  \item{\code{CONC}}{Plasma concentration of CP1805 (in ug/mL)}
  \item{\code{TIME}}{Time since first dose on \code{DAY} 1 (in hours)}
}}
\usage{
CP1805.conc
}
\description{
De-identified data set of concentration-time data for CP1805.
}
\seealso{
\code{\link{CP1805.dose}}, \code{\link{CP1805.demog}}
}
\keyword{datasets}

