% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP1805.R
\docType{data}
\name{CP1805.demog}
\alias{CP1805.demog}
\title{Demographics associated with the CP1805 datasets}
\format{A data frame with three variables:
\describe{
  \item{\code{ID}}{Subject identifier}
  \item{\code{SEX}}{Subject's sex (Male/Female)}
  \item{\code{WEIGHT}}{Subject's body weight (in kg)}
}}
\usage{
CP1805.demog
}
\description{
De-identified data set of demographic data for CP1805 pharmacokinetics.
}
\seealso{
\code{\link{CP1805.conc}}, \code{\link{CP1805.dose}}
}
\keyword{datasets}

