% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP1805.R
\docType{data}
\name{CP1805.dose}
\alias{CP1805.dose}
\title{Dosing of CP1805}
\format{A data frame with five variables:
\describe{
  \item{\code{ID}}{Subject identifier}
  \item{\code{ROUTE}}{Route of administration}
  \item{\code{DAY}}{Day of measurement}
  \item{\code{DOSE}}{Total dose of CP1805 (in mg)}
  \item{\code{DOSE.WEIGHT}}{Dose of CP1805 normalized by body weight (in mg/kg)}
  \item{\code{TIME.AFTER.DOSE}}{Time since the dose on the current \code{DAY} (in hours)}
  \item{\code{TIME}}{Time since first dose on \code{DAY} 1 (in hours)}
}}
\usage{
CP1805.dose
}
\description{
De-identified data set of dosing data for CP1805.
}
\seealso{
\code{\link{CP1805.conc}}, \code{\link{CP1805.demog}}
}
\keyword{datasets}
