% Generated by roxygen2 (4.0.0): do not edit by hand
\name{guessDateFormat}
\alias{guessDateFormat}
\title{Determine Format of Date and Date-Time Variables}
\usage{
guessDateFormat(x, returnDates = FALSE, tz = "")
}
\arguments{
\item{x}{character vector of dates or date-times}

\item{returnDates}{logical value; determines if POSIXlt objects are returned
or just the guessed format}

\item{tz}{character string; specifies the time zone to be used for the
conversion.  Defaults to the current time zone.}
}
\value{
By default, returns a character string representing the format of the
date-time variables.  If \code{returnDates} is TRUE, a vector of POSIXlt
objects is returned.
}
\description{
Given a vector of dates or date-times, determine the format if one is used
consistently.  If inconsistentencies are found, the function will fail.
The format can be used to create POSIXlt objects, which is done
automatically if \code{returnDates} is set.  See \code{\link{strptime}} for
examples of format specifications.
}
\details{
\code{guessDateFormat} is rigid when determining the date format.  For
date-times it expects the date and time parts to be separated by a space.  It
does not recognize all format specifications, such as the AM/PM indicator.
The time part should have at least one colon to separate hours from minutes.
The date part may have any separator as non-numerical values are removed
before parsing.
}
\examples{
x <- c("2014-01-15", "20140202")
guessDateFormat(x)
guessDateFormat(x, TRUE)
}
\author{
Cole Beck
}

