% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrocortisone.sim.R
\docType{data}
\name{hydrocortisone.sim.dose}
\alias{hydrocortisone.sim.dose}
\title{Dosing of hydrocortisone}
\format{A data frame with three variables:
\describe{
  \item{\code{ID}}{Subject identifier}
  \item{\code{TIME}}{Time (in hours)}
  \item{\code{DOSE}}{Dose of hydrocortisone (in mg)}
  \item{\code{ROUTE}}{Route of administration}
}}
\usage{
hydrocortisone.sim.dose
}
\description{
Simulated data set of dosing data for hydrocortisone.
}
\seealso{
\code{\link{hydrocortisone.sim.conc}}
}
\keyword{datasets}
