% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KI20160914.R
\docType{data}
\name{KI20160914.demog}
\alias{KI20160914.demog}
\title{Demographics associated with the KI20160914 datasets}
\format{A data frame with the following variables:
\describe{
  \item{\code{ID}}{Subject identifier}
  \item{\code{AGE}}{Subject's age (in years)}
  \item{\code{SEX}}{Subject's sex (Male/Female)}
  \item{\code{WEIGHT}}{Subject's body weight (in kg)}
}}
\usage{
KI20160914.demog
}
\description{
De-identified data set of demographic data for KI20160914 pharmacokinetics. 
(Note that KI20160914 is a de-identified name for the drug.)
}
\seealso{
\code{\link{KI20160914.conc}}, \code{\link{KI20160914.dose}}
}
\keyword{datasets}

