% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KI20160914.R
\docType{data}
\name{KI20160914.dose}
\alias{KI20160914.dose}
\title{Dosing of KI20160914}
\format{A data frame with the following variables:
\describe{
  \item{\code{ID}}{Subject identifier}
  \item{\code{PERIOD}}{Period number for different visits in the study}
  \item{\code{TIME}}{Nominal time since first dose on \code{DAY} 1 in the current \code{PERIOD} (in hours)}
  \item{\code{TIME.FIRST.DOSE}}{Nominal time since first dose in \code{PERIOD} 1 (in hours)}
  \item{\code{DAY}}{Study day (since the beginning of \code{PERIOD} 1}
  \item{\code{DOSE}}{Dose of KI20160914 (in mg).}
  \item{\code{FORMULATION}}{Formulation number}
  \item{\code{ROUTE}}{Route of administration}
}}
\usage{
KI20160914.dose
}
\description{
De-identified data set of dosing data for KI20160914.  (Note that
KI20160914 is a de-identified name for the drug.)
}
\seealso{
\code{\link{KI20160914.conc}},
  \code{\link{KI20160914.demog}}
}
\keyword{datasets}
