% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xanthohumol-lipoic.R
\docType{data}
\name{lipoic.acid.conc}
\alias{lipoic.acid.conc}
\title{Pharmacokinetics of racemic and r-lipoic acid}
\format{
A data frame with four variables:
\describe{
  \item{\code{ID}}{Subject identifier}
  \item{\code{TIME}}{Time (in minutes)}
  \item{\code{CONC}}{Plasma concentration of the analyte (in ng/mL)}
  \item{\code{ANALYTE}}{The name of the analyte measured in the concentration column}
  \item{\code{Period}}{The period number}
}
}
\source{
https://ir.library.oregonstate.edu/xmlui/bitstream/handle/1957/39168/KarnprachaChanida2013.pdf?sequence=1
}
\usage{
lipoic.acid.conc
}
\description{
Data set of concentration-time data for racemic and r-lipoic acid.
}
\references{
Karnpracha, Chanida. (2013) Comparison of Pharmacokinetic Data Analysis with
Two Competing Pharmacokinetic Software Program. (Master's Thesis)
}
\seealso{
\code{\link{lipoic.acid.dose}}, \code{\link{lipoic.acid.nca}},
  \code{\link{lipoic.acid.demog}}
}
\keyword{datasets}
