% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xanthohumol-lipoic.R
\docType{data}
\name{lipoic.acid.demog}
\alias{lipoic.acid.demog}
\title{Demographics for subjects dosed with xanthohumol}
\format{
A data frame with two variables:
\describe{
  \item{\code{ID}}{Subject identifier}
  \item{\code{Sex}}{The subject's sex (F=female, M=male)}
  \item{\code{Age}}{The subject's age category}
}
}
\source{
https://ir.library.oregonstate.edu/xmlui/bitstream/handle/1957/39168/KarnprachaChanida2013.pdf?sequence=1
}
\usage{
lipoic.acid.demog
}
\description{
Demographics for subjects dosed with xanthohumol
}
\references{
Karnpracha, Chanida. (2013) Comparison of Pharmacokinetic Data Analysis with
Two Competing Pharmacokinetic Software Program. (Master's Thesis)
}
\seealso{
\code{\link{lipoic.acid.dose}}, \code{\link{lipoic.acid.conc}},
  \code{\link{lipoic.acid.nca}}
}
\keyword{datasets}
