% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xanthohumol-lipoic.R
\docType{data}
\name{lipoic.acid.dose}
\alias{lipoic.acid.dose}
\title{Dosing of racemic and r-lipoic acid}
\format{A data frame with six variables:
\describe{
  \item{\code{ID}}{Subject identifier}
  \item{\code{Period}}{The period number}
  \item{\code{ANALYTE}}{The name of the analyte measured in the concentration column}
  \item{\code{TIME}}{Time (in hours)}
  \item{\code{DOSE}}{Dose of lipoic acid (in mg)}
  \item{\code{ROUTE}}{Route of administration}
}}
\source{
https://ir.library.oregonstate.edu/xmlui/bitstream/handle/1957/39168/KarnprachaChanida2013.pdf?sequence=1
}
\usage{
lipoic.acid.dose
}
\description{
Dosing of racemic and r-lipoic acid
}
\references{
Karnpracha, Chanida. (2013) Comparison of Pharmacokinetic Data Analysis with
Two Competing Pharmacokinetic Software Program. (Master's Thesis)
}
\seealso{
\code{\link{lipoic.acid.conc}}, \code{\link{lipoic.acid.nca}},
  \code{\link{lipoic.acid.demog}}
}
\keyword{datasets}

