% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xanthohumol-lipoic.R
\docType{data}
\name{lipoic.acid.nca}
\alias{lipoic.acid.nca}
\title{Noncompartmental analysis (NCA) parameters for racemic and r-lipoic acid as
calculated by Kinetica and WinNonlin.}
\format{
A data frame with five variables:
\describe{
  \item{\code{ID}}{Subject identifier}
  \item{\code{ANALYTE}}{The name of the analyte measured in the \code{Value} column}
  \item{\code{Parameter}}{NCA parameter name}
  \item{\code{Software}}{Name and version of software used for NCA calculation}
  \item{\code{Value}}{Value of the NCA parameter}
}
}
\source{
https://ir.library.oregonstate.edu/xmlui/bitstream/handle/1957/39168/KarnprachaChanida2013.pdf?sequence=1
}
\usage{
lipoic.acid.nca
}
\description{
Noncompartmental analysis (NCA) parameters for racemic and r-lipoic acid as
calculated by Kinetica and WinNonlin.
}
\references{
Karnpracha, Chanida. (2013) Comparison of Pharmacokinetic Data Analysis with
Two Competing Pharmacokinetic Software Program. (Master's Thesis)
}
\seealso{
\code{\link{lipoic.acid.dose}}, \code{\link{lipoic.acid.conc}},
  \code{\link{lipoic.acid.demog}}
}
\keyword{datasets}
