% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xanthohumol-lipoic.R
\docType{data}
\name{lipoic_acid_demog}
\alias{lipoic_acid_demog}
\title{Demographics for subjects dosed with r-lipoic acid}
\format{
A data frame with two variables:
\describe{
  \item{\code{ID}}{Subject identifier}
  \item{\code{Sex}}{The subject's sex (F=female, M=male)}
  \item{\code{Age}}{The subject's age category}
}
}
\source{
https://ir.library.oregonstate.edu/xmlui/bitstream/handle/1957/39168/KarnprachaChanida2013.pdf?sequence=1
}
\usage{
lipoic_acid_demog
}
\description{
Demographics for subjects dosed with r-lipoic acid
}
\references{
Karnpracha, Chanida. (2013) Comparison of Pharmacokinetic Data Analysis with
Two Competing Pharmacokinetic Software Program. (Master's Thesis)
}
\seealso{
Other lipoic acid concentration and NCA data: 
\code{\link{lipoic_acid_conc}},
\code{\link{lipoic_acid_dose}},
\code{\link{lipoic_acid_nca}}
}
\concept{lipoic acid concentration and NCA data}
\keyword{datasets}
