% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xanthohumol-lipoic.R
\docType{data}
\name{lipoic_acid_dose}
\alias{lipoic_acid_dose}
\title{Dosing of racemic and r-lipoic acid}
\format{
A data frame with six variables:
\describe{
  \item{\code{ID}}{Subject identifier}
  \item{\code{Period}}{The period number}
  \item{\code{ANALYTE}}{The name of the analyte measured in the concentration column}
  \item{\code{TIME}}{Time (in hours)}
  \item{\code{DOSE}}{Dose of lipoic acid (in mg)}
  \item{\code{ROUTE}}{Route of administration}
}
}
\source{
https://ir.library.oregonstate.edu/xmlui/bitstream/handle/1957/39168/KarnprachaChanida2013.pdf?sequence=1
}
\usage{
lipoic_acid_dose
}
\description{
Dosing of racemic and r-lipoic acid
}
\references{
Karnpracha, Chanida. (2013) Comparison of Pharmacokinetic Data Analysis with
Two Competing Pharmacokinetic Software Program. (Master's Thesis)
}
\seealso{
Other lipoic acid concentration and NCA data: 
\code{\link{lipoic_acid_conc}},
\code{\link{lipoic_acid_demog}},
\code{\link{lipoic_acid_nca}}
}
\concept{lipoic acid concentration and NCA data}
\keyword{datasets}
