% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xanthohumol-lipoic.R
\docType{data}
\name{lipoic_acid_nca}
\alias{lipoic_acid_nca}
\title{Noncompartmental analysis (NCA) parameters for racemic and r-lipoic acid as
calculated by Kinetica and WinNonlin.}
\format{
A data frame with five variables:
\describe{
  \item{\code{ID}}{Subject identifier}
  \item{\code{ANALYTE}}{The name of the analyte measured in the \code{Value} column}
  \item{\code{Parameter}}{NCA parameter name}
  \item{\code{Software}}{Name and version of software used for NCA calculation}
  \item{\code{Value}}{Value of the NCA parameter}
}
}
\source{
https://ir.library.oregonstate.edu/xmlui/bitstream/handle/1957/39168/KarnprachaChanida2013.pdf?sequence=1
}
\usage{
lipoic_acid_nca
}
\description{
Noncompartmental analysis (NCA) parameters for racemic and r-lipoic acid as
calculated by Kinetica and WinNonlin.
}
\references{
Karnpracha, Chanida. (2013) Comparison of Pharmacokinetic Data Analysis with
Two Competing Pharmacokinetic Software Program. (Master's Thesis)
}
\seealso{
Other lipoic acid concentration and NCA data: 
\code{\link{lipoic_acid_conc}},
\code{\link{lipoic_acid_demog}},
\code{\link{lipoic_acid_dose}}
}
\concept{lipoic acid concentration and NCA data}
\keyword{datasets}
