% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkdata-package.R
\name{pkdata-internal}
\alias{pkdata-internal}
\alias{.onLoad}
\title{Internal objects.}
\usage{
.onLoad(...)
}
\description{
Internal objects.
}
\details{
This function should not be called by the user.
}
\keyword{internal}
