% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkdata-package.R
\docType{package}
\name{pkdata-package}
\alias{pkdata-package}
\alias{pkdata}
\title{PK Data}
\description{
This package will prepare data sets for PK data analysis.
}
\author{
Cole Beck, Leena Choi

Maintainer: Cole Beck \email{cole.beck@vumc.org}
}
\keyword{package}
