% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortDoses.R
\name{sortDoses}
\alias{sortDoses}
\title{Sort Dose Data Set}
\usage{
sortDoses(
  data,
  idVar = "id",
  infusionDoseTimeVar = NULL,
  bolusDoseTimeVar = NULL,
  otherDoseTimeVar = NULL
)
}
\arguments{
\item{data}{data.frame; data set with dose information}

\item{idVar}{character string; name of ID variable, defaults to id}

\item{infusionDoseTimeVar}{character string; name of infusion dose time
variable}

\item{bolusDoseTimeVar}{character string; name of bolus dose time variable}

\item{otherDoseTimeVar}{character string; name of additional dose time
variable}
}
\value{
data.frame, containing dose data
}
\description{
Helper function to sort dose data set by ID and time variables.
}
\details{
Sort order will be the ID variable, then infusion time, bolus time, or other
dose date-time variables.
}
\author{
Cole Beck
}
