% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TC20160506.R
\docType{data}
\name{TC20160506.conc}
\alias{TC20160506.conc}
\title{Pharmacokinetics of TC20160506}
\format{A data frame with four variables:
\describe{
  \item{\code{ID}}{Subject identifier}
  \item{\code{ANALYTE}}{Analyte being measured}
  \item{\code{TIME}}{Time since first dose on \code{DAY} 1 (in hours)}
  \item{\code{CONC}}{Plasma concentration of TC20160506 (in ug/L)}
}}
\usage{
TC20160506.conc
}
\description{
De-identified data set of concentration-time data for TC20160506.  (Note that
TC20160506 is a de-identified name for the drug.)
}
\seealso{
\code{\link{TC20160506.dose}}, \code{\link{TC20160506.demog}}
}
\keyword{datasets}

