% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TC20160506.R
\docType{data}
\name{TC20160506.demog}
\alias{TC20160506.demog}
\title{Demographics associated with the TC20160506 datasets}
\format{A data frame with three variables:
\describe{
  \item{\code{ID}}{Subject identifier}
  \item{\code{AGE}}{Subject's age (in years)}
  \item{\code{SEX}}{Subject's sex (Male/Female)}
  \item{\code{WEIGHT}}{Subject's body weight (in kg)}
}}
\usage{
TC20160506.demog
}
\description{
De-identified data set of demographic data for TC20160506 pharmacokinetics. 
(Note that TC20160506 is a de-identified name for the drug.)
}
\seealso{
\code{\link{TC20160506.conc}}, \code{\link{TC20160506.dose}}
}
\keyword{datasets}

