% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TC20160506.R
\docType{data}
\name{TC20160506.dose}
\alias{TC20160506.dose}
\title{Dosing of TC20160506}
\format{A data frame with five variables:
\describe{
  \item{\code{ID}}{Subject identifier}
  \item{\code{ANALYTE}}{Analyte being dosed (the drug name)}
  \item{\code{TIME}}{Time since first dose on \code{DAY} 1 (in
    hours)}
  \item{\code{DOSE}}{Dose of TC20160506 (in ug).  Note that there
    may be more than one dose record per subject.}
  \item{\code{ROUTE}}{Route of administration}
}}
\usage{
TC20160506.dose
}
\description{
De-identified data set of dosing data for TC20160506.  (Note that
TC20160506 is a de-identified name for the drug.)
}
\seealso{
\code{\link{TC20160506.conc}},
  \code{\link{TC20160506.demog}}
}
\keyword{datasets}
