% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xanthohumol-lipoic.R
\docType{data}
\name{xanthohumol.demog}
\alias{xanthohumol.demog}
\title{Demographics for subjects dosed with xanthohumol}
\format{A data frame with two variables:
\describe{
  \item{\code{ID}}{Subject identifier}
  \item{\code{Sex}}{The subject's sex (F=female, M=male)}
}}
\usage{
xanthohumol.demog
}
\description{
Demographics for subjects dosed with xanthohumol
}
\keyword{datasets}

