% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xanthohumol-lipoic.R
\docType{data}
\name{xanthohumol.low.dose}
\alias{xanthohumol.low.dose}
\title{Dosing of xanthohumol}
\format{A data frame with four variables:
\describe{
  \item{\code{ID}}{Subject identifier}
  \item{\code{TIME}}{Time (in hours)}
  \item{\code{DOSE}}{Dose of xanthohumol (in mg)}
  \item{\code{ROUTE}}{Route of administration}
}}
\source{
https://ir.library.oregonstate.edu/xmlui/bitstream/handle/1957/39168/KarnprachaChanida2013.pdf?sequence=1
}
\usage{
xanthohumol.low.dose
}
\description{
Dosing of xanthohumol
}
\references{
Karnpracha, Chanida. (2013) Comparison of Pharmacokinetic Data Analysis with
Two Competing Pharmacokinetic Software Program. (Master's Thesis)
}
\seealso{
\code{\link{xanthohumol.low.conc}},
  \code{\link{xanthohumol.medium.conc}}, \code{\link{xanthohumol.high.conc}},
  \code{\link{xanthohumol.low.nca}}, \code{\link{xanthohumol.medium.nca}},
  \code{\link{xanthohumol.high.nca}}
}
\keyword{datasets}

