% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xanthohumol-lipoic.R
\docType{data}
\name{xanthohumol.low.nca}
\alias{xanthohumol.low.nca}
\title{Noncompartmental analysis (NCA) parameters for xanthohumol and isoxanthohumol
as calculated by Kinetica WinNonlin.}
\format{A data frame with five variables:
\describe{
  \item{\code{ID}}{Subject identifier}
  \item{\code{ANALYTE}}{The name of the analyte measured in the \code{Value} column}
  \item{\code{Parameter}}{NCA parameter name}
  \item{\code{Software}}{Name and version of software used for NCA calculation}
  \item{\code{Value}}{Value of the NCA parameter}
}}
\source{
https://ir.library.oregonstate.edu/xmlui/bitstream/handle/1957/39168/KarnprachaChanida2013.pdf?sequence=1
}
\usage{
xanthohumol.low.nca
}
\description{
Noncompartmental analysis (NCA) parameters for xanthohumol and isoxanthohumol
as calculated by Kinetica WinNonlin.
}
\references{
Karnpracha, Chanida. (2013) Comparison of Pharmacokinetic Data Analysis with
Two Competing Pharmacokinetic Software Program. (Master's Thesis)
}
\seealso{
\code{\link{xanthohumol.low.conc}}, 
  \code{\link{xanthohumol.medium.conc}},
  \code{\link{xanthohumol.high.conc}}, \code{\link{xanthohumol.low.dose}},
  \code{\link{xanthohumol.medium.dose}}, \code{\link{xanthohumol.high.dose}}
}
\keyword{datasets}

