% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xanthohumol-lipoic.R
\docType{data}
\name{xanthohumol_conc}
\alias{xanthohumol_conc}
\title{Data set of concentration-time data for xanthohumol and isoxanthohumol (a
metabolite of xanthohumol).}
\format{
A data frame with four variables:
\describe{
  \item{\code{ID}}{Subject identifier}
  \item{\code{TIME}}{Time (in hours)}
  \item{\code{CONC}}{Plasma concentration of the analyte (in ug/L)}
  \item{\code{ANALYTE}}{The name of the analyte measured in the concentration column}
}
}
\source{
https://ir.library.oregonstate.edu/xmlui/bitstream/handle/1957/39168/KarnprachaChanida2013.pdf?sequence=1
}
\usage{
xanthohumol_conc
}
\description{
Data set of concentration-time data for xanthohumol and isoxanthohumol (a
metabolite of xanthohumol).
}
\references{
Karnpracha, Chanida. (2013) Comparison of Pharmacokinetic Data Analysis with
Two Competing Pharmacokinetic Software Program. (Master's Thesis)
}
\seealso{
Other xanthohumol and isoxanthohumol concentration and NCA data: 
\code{\link{xanthohumol_dose}},
\code{\link{xanthohumol_nca}}
}
\concept{xanthohumol and isoxanthohumol concentration and NCA data}
\keyword{datasets}
