% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xanthohumol-lipoic.R
\docType{data}
\name{xanthohumol_dose}
\alias{xanthohumol_dose}
\title{Dosing of xanthohumol}
\format{
A data frame with four variables:
\describe{
  \item{\code{ID}}{Subject identifier}
  \item{\code{TIME}}{Time (in hours)}
  \item{\code{DOSE}}{Dose of xanthohumol (in mg)}
  \item{\code{ROUTE}}{Route of administration}
}
}
\source{
https://ir.library.oregonstate.edu/xmlui/bitstream/handle/1957/39168/KarnprachaChanida2013.pdf?sequence=1
}
\usage{
xanthohumol_dose
}
\description{
Dosing of xanthohumol
}
\references{
Karnpracha, Chanida. (2013) Comparison of Pharmacokinetic Data Analysis with
Two Competing Pharmacokinetic Software Program. (Master's Thesis)
}
\seealso{
Other xanthohumol and isoxanthohumol concentration and NCA data: 
\code{\link{xanthohumol_conc}},
\code{\link{xanthohumol_nca}}
}
\concept{xanthohumol and isoxanthohumol concentration and NCA data}
\keyword{datasets}
