% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP1805.R
\docType{data}
\name{CP1805}
\alias{CP1805}
\alias{CP1805_conc}
\alias{CP1805_dose}
\alias{CP1805_demog}
\title{Pharmacokinetics of CP1805}
\format{\code{CP1805_conc} is a de-identified data set of concentration-time
  data for CP1805 as a data frame with the following variables:
\describe{
  \item{\code{ID}}{Subject identifier}
  \item{\code{DAY}}{Day of measurement}
  \item{\code{TIME_AFTER_DOSE}}{Time since the dose on the current \code{DAY} (in hours)}
  \item{\code{TIME}}{Time since first dose on \code{DAY} 1 (in hours)}
  \item{\code{TIMEU}}{Units for time, "hours"}
  \item{\code{CONC}}{Plasma concentration of CP1805 (in ug/mL)}
  \item{\code{CONCU}}{Units for plasma concentration of CP1805, "ug/mL"}
}}
\usage{
CP1805_conc

CP1805_dose

CP1805_demog
}
\description{
Pharmacokinetics of CP1805
}
\keyword{datasets}
