% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorouracil.R
\docType{data}
\name{fluorouracil}
\alias{fluorouracil}
\alias{fluorouracil_conc}
\alias{fluorouracil_dose}
\alias{fluorouracil_demog}
\title{Pharmacokinetics of fluorouracil}
\format{\code{fluorouracil_conc} is a data frame describing concentration by time with the following variables:
\describe{
  \item{\code{ID}}{Subject identifier}
  \item{\code{TIME}}{Time (in minutes)}
  \item{\code{TIMEU}}{Units for time, "minutes"}
  \item{\code{NDOSE}}{Nominal dose (in mg/kg)}
  \item{\code{NDOSEU}}{Units for nominal dose, "mg/kg"}
  \item{\code{CONC}}{Plasma concentration of fluorouracil (in ug/mL)}
  \item{\code{CONCU}}{Units for plasma concentration of fluorouracil, "ug/mL"}
}}
\usage{
fluorouracil_conc

fluorouracil_dose

fluorouracil_demog
}
\description{
Observed data set of concentration-time data for fluorouracil.
}
\references{
MacMillan, William E., William H. Wolberg, and Peter G. Welling. 1978.
"Pharmacokinetics of Fluorouracil in Humans." Cancer Research 38 (10):
3479-82.
}
\keyword{datasets}
