% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TC20160506.R
\docType{data}
\name{TC20160506}
\alias{TC20160506}
\alias{TC20160506_conc}
\alias{TC20160506_dose}
\alias{TC20160506_demog}
\title{Pharmacokinetics of TC20160506 (Note that TC20160506 is a de-identified name for the drug.)}
\format{\code{TC20160506_conc} is concentration-time data.  The data.frame
  has the following variables:
\describe{
  \item{\code{ID}}{Subject identifier}
  \item{\code{ANALYTE}}{Analyte being measured}
  \item{\code{TIME}}{Time since first dose on \code{DAY} 1 (in hours)}
  \item{\code{TIMEU}}{Units for time, "hours"}
  \item{\code{CONC}}{Plasma concentration of TC20160506 (in ug/L)}
  \item{\code{CONCU}}{Units for plasma concentration of TC20160506, "ug/L"}
}}
\usage{
TC20160506_conc

TC20160506_dose

TC20160506_demog
}
\description{
De-identified data set of concentration-time data for TC20160506. When
working with this dataset, be aware that the \code{ANALYTE} named "Parent"
can be converted into "Metabolite".
}
\keyword{datasets}
